backend/requirements.txt
django
djangorestframework
django-cors-headers
django-rest-auth
psycopg2-binary
gunicorn